requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.5
cycler==0.11.0
docker-pycreds==0.4.0
fonttools==4.37.4
gitdb==4.0.9
GitPython==3.1.29
idna==3.4
kiwisolver==1.4.4
matplotlib==3.6.1
numpy==1.23.3
packaging==21.3
pandas==1.5.0
pathtools==0.1.2
Pillow==9.2.0
promise==2.3
protobuf==4.21.8
psutil==5.9.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-slugify==6.1.2
pytz==2022.4
PyYAML==6.0
requests==2.28.1
sentry-sdk==1.9.10
setproctitle==1.3.2
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
text-unidecode==1.3
tqdm==4.64.1
urllib3==1.26.12
wandb==0.13.4
